Rename `proxy` to `proxy_socks` to pave way for `proxy_ssh`
