+++ lib/proxy_socks/proxy.go
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-package proxy
+package proxy_socks
 
 import (
 	"net"
